refactor(core): 検索クエリ計画を search/query_plan.rs へ分割 (#599)
`prepare_query_plan` は 1 検索呼び出しあたりのクエリ導出（正規化・dot/path 判定・
Fuzzy bitmask・migemo かなクエリ・UTF-32 needle・パス照合クエリ・履歴キー）を
純粋に行う自由関数で、エントリ走査も履歴更新もしない。検索オーケストレーションから
独立させ、製品挙動は変えずに private submodule へ移設する。

- 移設: `QueryPlan` と `prepare_query_plan`
- `QueryPlan` とフィールドは `pub(super)` で親 search のみに公開（crate 可視性は広げない）
- 正準は `crate::query` の正規化群（`normalize_query` / `normalize_history_query_key` /
  `char_bitmask` / `to_kana`）のまま。`kana_char_mask` は search.rs 残置を継続共有
- `decide_incremental` / `prev_*` の read/write は search.rs に残す
- move で不要化した search.rs の import 5 つを除去（crate::query 群 4 + `std::borrow::Cow`）
- snotra-core/CLAUDE.md の search.rs 索引に query_plan.rs を追記

不変条件: ASCII 小文字の Cow 借用経路・needle_u32 の 1 回生成・has_dot 意味・path_query
の生クエリ保持・path_history_key と path_query の非混同・区切り判定・kana ASCII 残留棄却。

検証: cargo test -p snotra-core（466 passed, 9 ignored。受け入れ 4 テスト
incremental_kana_min_chars_crossing / _ascii_residue_clearing / path_match_consecutive_spaces /
path_match_history_key_unified を含む）/ clippy / rustdoc(-D warnings) /
cargo check -p snotra（下流）/ governance:check G1..G9。
