+++ snotra-core/CLAUDE.md
@@ -28,6 +28,7 @@
   - **`kana_lower_names` / `kana_char_masks` は `migemo_enabled` が true のときのみ構築し、無効時は空 Vec**（migemo 無効ユーザーの死蔵メモリ ~2.1–2.7MB/50k を削る・構築も約 2 倍速、issue #337）。2 つの kana 系 Vec は必ず同時に空/同長（`assemble` の debug_assert が検証）。空 Vec のとき検索ループは `kana_available` 空ガードで `kana_lower_names[i]` アクセスを回避し、Fuzzy pre-filter は `kana_char_masks.is_empty()` チェックで kana 経路を棄却する（構築時 migemo OFF→検索時 ON の窓での panic 防止）
   - **migemo トグルの反映は index 再構築経由**: `update_config` は engine を再構築しないため、`config_watcher` が engine の `IndexInputs` 差分で `start_index_build` を kick する再構築に依存する（#347 Phase 2 で `needs_reindex` は `IndexInputs` に統合）
   - **パスマッチング**: クエリにパス区切り文字（`\` `/`）を含む場合、`normalized_key`（= `normalize_entry_key(target_path)`）に対して Substring マッチを試みる。スコアは `3000 - min(byte_pos, 500)`。name/file_name/kana 全て不成立時のフォールバック。`has_path_sep` 時は Fuzzy ビットマスク pre-filter をスキップする
+  - **クエリ計画は `search/query_plan.rs` に分離**（#599。責務は `//!`）: `QueryPlan` と `prepare_query_plan`（正規化クエリ・dot/path 判定・Fuzzy bitmask・migemo かなクエリ・UTF-32 needle・パス照合クエリ・履歴キーの純粋導出）。`decide_incremental` / `prev_*` の read/write は `search.rs` に残す。`QueryPlan` とフィールドは `pub(super)` で親のみに公開
   - **構築処理は `search/build.rs` に分離**（#598。責務は `//!`）: Wave 1/2・kana マスクの並列構築、IndexCache 復元（v4 ヒット時 Wave 1 スキップ / v3 fallback）、全コンストラクタ（`new` / `new_with_migemo` / `new_with_cached_masks` / `assemble`）。検索ホットパスは `search.rs` に残す。`kana_char_mask`（query 側と共有しうる純粋関数）は `search.rs` 側に残置
   - **ユニットテストは `search/tests/` に機能別分割**（#597。責務は各ファイルの `//!`・製品コードは `search.rs` のまま）: 索引 `search/tests/mod.rs`、共通 fixture `search/tests/common.rs`、`search/tests/basic.rs`（基本検索・拡張子・正規化）/ `search/tests/ranking.rs`（top-k・タイブレーク・ビットマスク）/ `search/tests/incremental.rs`（incremental キャッシュ）/ `search/tests/migemo.rs`（かな検索・条件付き構築）/ `search/tests/path.rs`（パスマッチ）/ `search/tests/performance.rs`（`#[ignore]` ベンチ・メモリ計測）
 - `history.rs`: 起動履歴・クエリ別履歴・フォルダ展開履歴の管理、バイナリ永続化
